+++ CHANGELOG.md
@@ -9,6 +9,7 @@
 - COD10-013: Add location filter and weekday-colored hourly chart to Visits bar stats (#89) [424]
 - COD10-016: Add mode filter on pie stats and calendar view for Travels (#91) [399]
 - COD10-017: Autocomplete travel destination from origin + time-of-day history (#92) [114]
+- COD10-040: Category cards + generic subdivision registry (CABA barrios) on Locations (#93) [2367]
 
 ## v0.9.0
 - COD9-003: Add Partido for Locations with B-Prefix zipcodes (#57) [388]